Rearranged basic layout
