+++ src/components/About/index.js
@@ -1,5 +1,5 @@
 // src/components/About/index.js
-import React, { PropTypes, Component } from 'react';
+import React, { Component } from 'react';
 import classnames from 'classnames';
 
 import './style.css';
