+++ src/components/App/index.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import Navigation from './Navigation';
+import './index.css';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <Navigation />
+        <div className="App-main">
+          {this.props.children}
+        </div>
+      </div>
+    );
+  }
+}
+
+export default App;
