+++ src/components/Index/index.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Index extends Component {
+  render() {
+    return (
+      <div>
+        <h1>Index Page</h1>
+      </div>
+    );
+  }
+}
+
+export default Index;
