+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/Main/App';
 import Routes from './routes';
 import { browserHistory } from 'react-router';
